pyproject.toml
[project]
name = "ai-health-example"
version = "0.1.0"
description = "AI 헬스케어 파이널 프로젝트 예시용"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "cryptography>=46.0.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "redis>=7.1.0",
    "tomlkit>=0.14.0",
]

[dependency-groups]
ai = [
    "scikit-learn>=1.8.0",
    "torch",
    "torchvision",
    "torchaudio",
    "sentence-transformers>=5.2.0"
]
app = [
    "aerich>=0.9.2",
    "asyncpg>=0.30.0",
    "bcrypt<=4.0.1",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "orjson>=3.11.5",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "tortoise-orm>=0.25.3",
    "uvicorn>=0.40.0",
]
dev = [
    "coverage>=7.13.2",
    "mypy>=1.19.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "types-passlib>=1.7.7.20250602",
    "types-python-dateutil>=2.9.0.20260124",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.aerich]
tortoise_orm = "app.db.databases.TORTOISE_ORM"
location = "./app/db/migrations"
src_folder = "./."

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
# 기본 라인 길이 설정
line-length = 120
target-version = "py312"

[tool.ruff.lint]
# 사용할 린트 규칙들 선택
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C90", # mccabe
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
]
ignore = ["UP046", "E501"]

# 특정 파일에서 무시할 규칙들
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"db/migrations/*" = ["ALL"]

[tool.ruff.format]
# 포매터 설정
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
